Don't dynamically change config mid-Kafka test
